+++ changelog.md
@@ -1,11 +1,31 @@
-#Change Log
+# Change Log
 
 Team membership:  Teresa (Captain) & Jayden (Mate)  
 Team conventions: Allman notation, markdown for changelog  
 Changelog format: [Markdown](https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet) 
 
-## *Version 1.0*
+## *Version 1.0.1*
 
+### 2017-09-23. (Teresa)
+- Added merged develop branch with team repo.
 
-[1.0.0] - 2017-09-14
-This is our initial additional to our changelog.
+## *Version 1.0.0*
+
+### 2017-09-22. (Jayden)
+- Improvement on templates
+  - generate content for images using a table cell template.
+
+### 2017-09-22. (Teresa)
+- Updated Base Controller
+- Updated 4 views:
+  - template.php
+  - about.php
+  - welcome.php
+  - gallery.php
+
+### 2017-09-21. (Jayden)
+- Added two pages:
+  - About
+  - Gallery
+- Added CodeIgniter framework to the project.
+- This is our initial additional to our changelog.